.NET 6.0

+++ src/Catharsis.Commons/FileSystemExtensions.cs
@@ -16,7 +16,7 @@ public static class FileSystemExtensions
   /// </summary>
   /// <param name="drive"></param>
   /// <returns></returns>
-  public static bool IsEmpty(this DriveInfo drive) => drive.RootDirectory.IsEmpty();
+  public static bool IsEmpty(this DriveInfo drive) => drive is not null ? drive.RootDirectory.IsEmpty() : throw new ArgumentNullException(nameof(drive));
 
   /// <summary>
   ///   <para></para>
@@ -30,7 +30,7 @@ public static class FileSystemExtensions
   /// </summary>
   /// <param name="file"></param>
   /// <returns></returns>
-  public static bool IsEmpty(this FileInfo file) => !file.Exists || file.Length <= 0;
+  public static bool IsEmpty(this FileInfo file) => file is not null ? !file.Exists || file.Length <= 0 : throw new ArgumentNullException(nameof(file));
 
   /// <summary>
   ///   <para>Erases all content from a file, making it a zero-length one.</para>
@@ -39,7 +39,10 @@ public static class FileSystemExtensions
   /// <returns>Back reference to the current file.</returns>
   public static FileInfo Empty(this FileInfo file)
   {
+    if (file is null) throw new ArgumentNullException(nameof(file));
+
     file.CreateWithPath().Refresh();
+
     return file;
   }
 
@@ -50,6 +53,8 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static DirectoryInfo Empty(this DirectoryInfo directory)
   {
+    if (directory is null) throw new ArgumentNullException(nameof(directory));
+
     directory.Directories().ForEach(directory => directory.Delete(true));
     directory.Files().ForEach(file => file.Delete());
 
@@ -63,6 +68,8 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static FileInfo CreateWithPath(this FileInfo file)
   {
+    if (file is null) throw new ArgumentNullException(nameof(file));
+
     if (file.DirectoryName != null)
     {
       Directory.CreateDirectory(file.DirectoryName);
@@ -81,7 +88,10 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static IEnumerable<string> Lines(this FileInfo file, Encoding encoding = null)
   {
+    if (file is null) throw new ArgumentNullException(nameof(file));
+
     using var reader = file.ToStreamReader(encoding);
+
     return reader.Lines();
   }
 
@@ -93,6 +103,8 @@ public static class FileSystemExtensions
   /// <returns>List of strings which have been read from a <paramref name="file"/>.</returns>
   public static async IAsyncEnumerable<string> LinesAsync(this FileInfo file, Encoding encoding = null)
   {
+    if (file is null) throw new ArgumentNullException(nameof(file));
+
     using var reader = file.ToStreamReader(encoding);
 
     await foreach (var line in reader.LinesAsync().ConfigureAwait(false))
@@ -111,6 +123,9 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static T Print<T>(this T instance, FileInfo destination, Encoding encoding = null)
   {
+    if (instance is null) throw new ArgumentNullException(nameof(instance));
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+
     using var writer = destination.ToStreamWriter(encoding);
 
     return instance.Print(writer);
@@ -126,6 +141,9 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static async Task<T> PrintAsync<T>(this T instance, FileInfo destination, Encoding encoding = null, CancellationToken cancellation = default)
   {
+    if (instance is null) throw new ArgumentNullException(nameof(instance));
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+
     await using var writer = destination.ToStreamWriter(encoding);
 
     return await instance.PrintAsync(writer, cancellation).ConfigureAwait(false);
@@ -137,14 +155,11 @@ public static class FileSystemExtensions
   /// <param name="file"></param>
   /// <param name="action"></param>
   /// <returns></returns>
-  public static FileInfo TryFinallyClear(this FileInfo file, Action<FileInfo> action)
-  {
-    return file.TryFinally(file =>
+  public static FileInfo TryFinallyClear(this FileInfo file, Action<FileInfo> action) => file.TryFinally(file =>
     {
       file.CreateWithPath();
       action(file);
     }, file => file.Empty());
-  }
 
   /// <summary>
   ///   <para></para>
@@ -152,14 +167,11 @@ public static class FileSystemExtensions
   /// <param name="directory"></param>
   /// <param name="action"></param>
   /// <returns></returns>
-  public static DirectoryInfo TryFinallyClear(this DirectoryInfo directory, Action<DirectoryInfo> action)
-  {
-    return directory.TryFinally(directory =>
+  public static DirectoryInfo TryFinallyClear(this DirectoryInfo directory, Action<DirectoryInfo> action) => directory.TryFinally(directory =>
     {
       directory.Create();
       action(directory);
     }, directory => directory.Empty());
-  }
 
   /// <summary>
   ///   <para></para>
@@ -167,14 +179,11 @@ public static class FileSystemExtensions
   /// <param name="file"></param>
   /// <param name="action"></param>
   /// <returns></returns>
-  public static FileInfo TryFinallyDelete(this FileInfo file, Action<FileInfo> action)
-  {
-    return file.TryFinally(file =>
+  public static FileInfo TryFinallyDelete(this FileInfo file, Action<FileInfo> action) => file.TryFinally(file =>
     {
       file.CreateWithPath();
       action(file);
     }, file => file.Delete());
-  }
 
   /// <summary>
   ///   <para></para>
@@ -182,14 +191,11 @@ public static class FileSystemExtensions
   /// <param name="directory"></param>
   /// <param name="action"></param>
   /// <returns></returns>
-  public static DirectoryInfo TryFinallyDelete(this DirectoryInfo directory, Action<DirectoryInfo> action)
-  {
-    return directory.TryFinally(directory =>
+  public static DirectoryInfo TryFinallyDelete(this DirectoryInfo directory, Action<DirectoryInfo> action) => directory.TryFinally(directory =>
     {
       directory.Create();
       action(directory);
     }, directory => directory.Delete(true));
-  }
 
   /// <summary>
   ///   <para></para>
@@ -198,7 +204,7 @@ public static class FileSystemExtensions
   /// <param name="pattern"></param>
   /// <param name="recursive"></param>
   /// <returns></returns>
-  public static IEnumerable<FileInfo> Files(this DirectoryInfo directory, string pattern = null, bool recursive = false) => directory.Exists ? directory.EnumerateFiles(pattern ?? "*", new EnumerationOptions { RecurseSubdirectories = recursive }) : Enumerable.Empty<FileInfo>();
+  public static IEnumerable<FileInfo> Files(this DirectoryInfo directory, string pattern = null, bool recursive = false) => directory is not null ? directory.Exists ? directory.EnumerateFiles(pattern ?? "*", new EnumerationOptions { RecurseSubdirectories = recursive }) : Enumerable.Empty<FileInfo>() : throw new ArgumentNullException(nameof(directory));
 
   /// <summary>
   ///   <para></para>
@@ -207,7 +213,7 @@ public static class FileSystemExtensions
   /// <param name="pattern"></param>
   /// <param name="recursive"></param>
   /// <returns></returns>
-  public static IEnumerable<DirectoryInfo> Directories(this DriveInfo drive, string pattern = null, bool recursive = false) => drive.RootDirectory.Directories(pattern, recursive);
+  public static IEnumerable<DirectoryInfo> Directories(this DriveInfo drive, string pattern = null, bool recursive = false) => drive is not null ? drive.RootDirectory.Directories(pattern, recursive) : throw new ArgumentNullException(nameof(drive));
 
   /// <summary>
   ///   <para></para>
@@ -216,7 +222,7 @@ public static class FileSystemExtensions
   /// <param name="pattern"></param>
   /// <param name="recursive"></param>
   /// <returns></returns>
-  public static IEnumerable<DirectoryInfo> Directories(this DirectoryInfo directory, string pattern = null, bool recursive = false) => directory.Exists ? directory.EnumerateDirectories(pattern ?? "*", new EnumerationOptions { RecurseSubdirectories = recursive }) : Enumerable.Empty<DirectoryInfo>();
+  public static IEnumerable<DirectoryInfo> Directories(this DirectoryInfo directory, string pattern = null, bool recursive = false) => directory is not null ? directory.Exists ? directory.EnumerateDirectories(pattern ?? "*", new EnumerationOptions { RecurseSubdirectories = recursive }) : Enumerable.Empty<DirectoryInfo>() : throw new ArgumentNullException(nameof(directory));
 
   /// <summary>
   ///   <para></para>
@@ -225,7 +231,7 @@ public static class FileSystemExtensions
   /// <param name="pattern"></param>
   /// <param name="recursive"></param>
   /// <returns></returns>
-  public static long Size(this DriveInfo drive, string pattern = null, bool recursive = true) => drive.RootDirectory.Size(pattern, recursive);
+  public static long Size(this DriveInfo drive, string pattern = null, bool recursive = true) => drive is not null ? drive.RootDirectory.Size(pattern, recursive) : throw new ArgumentNullException(nameof(drive));
 
   /// <summary>
   ///   <para></para>
@@ -241,7 +247,7 @@ public static class FileSystemExtensions
   /// </summary>
   /// <param name="entry"></param>
   /// <returns></returns>
-  public static Uri ToUri(this FileSystemInfo entry) => new(entry.FullName);
+  public static Uri ToUri(this FileSystemInfo entry) => entry is not null ? new Uri(entry.FullName) : throw new ArgumentNullException(nameof(entry));
 
   /// <summary>
   ///   <para></para>
@@ -250,36 +256,28 @@ public static class FileSystemExtensions
   /// <param name="pattern"></param>
   /// <param name="recursive"></param>
   /// <returns></returns>
-  public static IEnumerable<FileSystemInfo> ToEnumerable(this DirectoryInfo directory, string pattern = null, bool recursive = false) => directory.Exists ? directory.EnumerateFileSystemInfos(pattern ?? "*", new EnumerationOptions { RecurseSubdirectories = recursive }) : Enumerable.Empty<FileSystemInfo>();
+  public static IEnumerable<FileSystemInfo> ToEnumerable(this DirectoryInfo directory, string pattern = null, bool recursive = false) => directory is not null ? directory.Exists ? directory.EnumerateFileSystemInfos(pattern ?? "*", new EnumerationOptions { RecurseSubdirectories = recursive }) : Enumerable.Empty<FileSystemInfo>() : throw new ArgumentNullException(nameof(directory));
 
   /// <summary>
   ///   <para></para>
   /// </summary>
   /// <param name="file"></param>
   /// <returns></returns>
-  public static FileStream ToStream(this FileInfo file) => file.Open(FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.None);
+  public static FileStream ToStream(this FileInfo file) => file is not null ? file.Open(FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.None) : throw new ArgumentNullException(nameof(file));
 
   /// <summary>
   ///   <para></para>
   /// </summary>
   /// <param name="file"></param>
   /// <returns></returns>
-  public static FileStream ToReadOnlyStream(this FileInfo file) => file.Open(FileMode.Open, FileAccess.Read, FileShare.Read);
+  public static FileStream ToReadOnlyStream(this FileInfo file) => file is not null ? file.Open(FileMode.Open, FileAccess.Read, FileShare.Read) : throw new ArgumentNullException(nameof(file));
 
   /// <summary>
   ///   <para></para>
   /// </summary>
   /// <param name="file"></param>
   /// <returns></returns>
-  public static FileStream ToWriteOnlyStream(this FileInfo file) => file.Open(FileMode.Append, FileAccess.Write, FileShare.None);
-
-  /// <summary>
-  ///   <para></para>
-  /// </summary>
-  /// <param name="file"></param>
-  /// <param name="encoding"></param>
-  /// <returns></returns>
-  public static StreamReader ToStreamReader(this FileInfo file, Encoding encoding = null) => new(file.FullName, encoding ?? Encoding.Default);
+  public static FileStream ToWriteOnlyStream(this FileInfo file) => file is not null ? file.Open(FileMode.Append, FileAccess.Write, FileShare.None) : throw new ArgumentNullException(nameof(file));
 
   /// <summary>
   ///   <para></para>
@@ -287,7 +285,15 @@ public static class FileSystemExtensions
   /// <param name="file"></param>
   /// <param name="encoding"></param>
   /// <returns></returns>
-  public static StreamWriter ToStreamWriter(this FileInfo file, Encoding encoding = null) => new(file.FullName, true, encoding ?? Encoding.Default, -1);
+  public static StreamReader ToStreamReader(this FileInfo file, Encoding encoding = null) => file is not null ? new StreamReader(file.FullName, encoding ?? Encoding.Default) : throw new ArgumentNullException(nameof(file));
+
+  /// <summary>
+  ///   <para></para>
+  /// </summary>
+  /// <param name="file"></param>
+  /// <param name="encoding"></param>
+  /// <returns></returns>
+  public static StreamWriter ToStreamWriter(this FileInfo file, Encoding encoding = null) => file is not null ? new StreamWriter(file.FullName, true, encoding ?? Encoding.Default, -1) : throw new ArgumentNullException(nameof(file));
 
   /// <summary>
   ///   <para></para>
@@ -296,7 +302,10 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static IEnumerable<byte> ToBytes(this FileInfo source)
   {
+    if (source is null) throw new ArgumentNullException(nameof(source));
+
     using var stream = source.ToReadOnlyStream();
+
     return stream.ToBytes();
   }
 
@@ -308,6 +317,8 @@ public static class FileSystemExtensions
   /// <returns>Byte content of specified <paramref name="source"/>.</returns>
   public static async IAsyncEnumerable<byte> ToBytesAsync(this FileInfo source, [EnumeratorCancellation] CancellationToken cancellation = default)
   {
+    if (source is null) throw new ArgumentNullException(nameof(source));
+
     await using var stream = source.ToReadOnlyStream();
 
     await foreach (var element in stream.ToBytesAsync(cancellation).ConfigureAwait(false))
@@ -324,7 +335,10 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static string ToText(this FileInfo source, Encoding encoding = null)
   {
+    if (source is null) throw new ArgumentNullException(nameof(source));
+
     using var reader = source.ToStreamReader(encoding);
+
     return reader.ToText();
   }
 
@@ -336,7 +350,10 @@ public static class FileSystemExtensions
   /// <returns>Text contents of a <paramref name="source"/>.</returns>
   public static async Task<string> ToTextAsync(this FileInfo source, Encoding encoding = null)
   {
+    if (source is null) throw new ArgumentNullException(nameof(source));
+
     using var reader = source.ToStreamReader(encoding);
+
     return await reader.ToTextAsync().ConfigureAwait(false);
   }
 
@@ -348,9 +365,13 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static FileInfo WriteBytes(this FileInfo destination, IEnumerable<byte> bytes)
   {
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+    if (bytes is null) throw new ArgumentNullException(nameof(bytes));
+
     try
     {
       using var stream = destination.ToWriteOnlyStream();
+
       stream.WriteBytes(bytes);
 
       return destination;
@@ -370,9 +391,13 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static async Task<FileInfo> WriteBytesAsync(this FileInfo destination, IEnumerable<byte> bytes, CancellationToken cancellation = default)
   {
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+    if (bytes is null) throw new ArgumentNullException(nameof(bytes));
+
     try
     {
       await using var stream = destination.ToWriteOnlyStream();
+
       await stream.WriteBytesAsync(bytes, cancellation).ConfigureAwait(false);
 
       return destination;
@@ -393,9 +418,13 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static async Task<FileInfo> WriteTextAsync(this FileInfo destination, string text, Encoding encoding = null, CancellationToken cancellation = default)
   {
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+    if (text is null) throw new ArgumentNullException(nameof(text));
+
     try
     {
       await using var writer = destination.ToStreamWriter(encoding);
+
       await writer.WriteTextAsync(text, cancellation).ConfigureAwait(false);
 
       return destination;
@@ -415,6 +444,9 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static FileInfo WriteText(this FileInfo destination, string text, Encoding encoding = null)
   {
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+    if (text is null) throw new ArgumentNullException(nameof(text));
+
     try
     {
       using var writer = destination.ToStreamWriter(encoding);
@@ -436,7 +468,11 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static IEnumerable<byte> WriteTo(this IEnumerable<byte> bytes, FileInfo destination)
   {
+    if (bytes is null) throw new ArgumentNullException(nameof(bytes));
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+
     destination.WriteBytes(bytes);
+
     return bytes;
   }
 
@@ -449,7 +485,11 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static async Task<IEnumerable<byte>> WriteToAsync(this IEnumerable<byte> bytes, FileInfo destination, CancellationToken cancellation = default)
   {
+    if (bytes is null) throw new ArgumentNullException(nameof(bytes));
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+
     await destination.WriteBytesAsync(bytes, cancellation).ConfigureAwait(false);
+
     return bytes;
   }
 
@@ -462,7 +502,11 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static string WriteTo(this string text, FileInfo destination, Encoding encoding = null)
   {
+    if (text is null) throw new ArgumentNullException(nameof(text));
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+
     destination.WriteText(text, encoding);
+
     return text;
   }
 
@@ -476,7 +520,11 @@ public static class FileSystemExtensions
   /// <returns></returns>
   public static async Task<string> WriteToAsync(this string text, FileInfo destination, Encoding encoding = null, CancellationToken cancellation = default)
   {
+    if (text is null) throw new ArgumentNullException(nameof(text));
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+
     await destination.WriteTextAsync(text, encoding, cancellation).ConfigureAwait(false);
+
     return text;
   }
 }
