+++ src/Catharsis.Commons/ProcessExtensions.cs
@@ -16,6 +16,8 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static Process Restart(this Process process, TimeSpan? timeout = null)
   {
+    if (process is null) throw new ArgumentNullException(nameof(process));
+
     process.Finish(timeout);
 
     var restarted = new Process();
@@ -33,6 +35,8 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static Process Finish(this Process process, TimeSpan? timeout = null)
   {
+    if (process is null) throw new ArgumentNullException(nameof(process));
+
     if (timeout != null)
     {
       if (!process.WaitForExit((int) timeout.Value.TotalMilliseconds))
@@ -57,8 +61,12 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static async Task<Process> FinishAsync(this Process process, CancellationToken cancellation = default)
   {
+    if (process is null) throw new ArgumentNullException(nameof(process));
+
     cancellation.Register(process.Kill);
+
     await process.WaitForExitAsync(cancellation).ConfigureAwait(false);
+
     return process;
   }
 #endif
@@ -94,21 +102,21 @@ public static class ProcessExtensions
   /// </summary>
   /// <param name="process"></param>
   /// <returns></returns>
-  public static IEnumerable<byte> ToBytes(this Process process) => process.StandardOutput.BaseStream.ToBytes();
+  public static IEnumerable<byte> ToBytes(this Process process) => process is not null ? process.StandardOutput.BaseStream.ToBytes() : throw new ArgumentNullException(nameof(process));
 
   /// <summary>
   ///   <para></para>
   /// </summary>
   /// <param name="process"></param>
   /// <returns></returns>
-  public static string ToText(this Process process) => process.StandardOutput.ToText();
+  public static string ToText(this Process process) => process is not null ? process.StandardOutput.ToText() : throw new ArgumentNullException(nameof(process));
 
   /// <summary>
   ///   <para></para>
   /// </summary>
   /// <param name="process"></param>
   /// <returns></returns>
-  public static string ToErrorText(this Process process) => process.StandardError.ToText();
+  public static string ToErrorText(this Process process) => process is not null ? process.StandardError.ToText() : throw new ArgumentNullException(nameof(process));
 
   /// <summary>
   ///   <para></para>
@@ -116,21 +124,21 @@ public static class ProcessExtensions
   /// <param name="process"></param>
   /// <param name="cancellation"></param>
   /// <returns></returns>
-  public static IAsyncEnumerable<byte> ToBytesAsync(this Process process, CancellationToken cancellation = default) => process.StandardOutput.BaseStream.ToBytesAsync(cancellation);
+  public static IAsyncEnumerable<byte> ToBytesAsync(this Process process, CancellationToken cancellation = default) => process is not null ? process.StandardOutput.BaseStream.ToBytesAsync(cancellation) : throw new ArgumentNullException(nameof(process));
 
   /// <summary>
   ///   <para></para>
   /// </summary>
   /// <param name="process"></param>
   /// <returns></returns>
-  public static async Task<string> ToTextAsync(this Process process) => await process.StandardOutput.ToTextAsync().ConfigureAwait(false);
+  public static async Task<string> ToTextAsync(this Process process) => process is not null ? await process.StandardOutput.ToTextAsync().ConfigureAwait(false) : throw new ArgumentNullException(nameof(process));
 
   /// <summary>
   ///   <para></para>
   /// </summary>
   /// <param name="process"></param>
   /// <returns></returns>
-  public static async Task<string> ToErrorTextAsync(this Process process) => await process.StandardError.ToTextAsync().ConfigureAwait(false);
+  public static async Task<string> ToErrorTextAsync(this Process process) => process is not null ? await process.StandardError.ToTextAsync().ConfigureAwait(false) : throw new ArgumentNullException(nameof(process));
 
   /// <summary>
   ///   <para></para>
@@ -140,7 +148,11 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static Process WriteBytes(this Process destination, IEnumerable<byte> bytes)
   {
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+    if (bytes is null) throw new ArgumentNullException(nameof(bytes));
+
     destination.StandardInput.BaseStream.WriteBytes(bytes);
+
     return destination;
   }
 
@@ -152,6 +164,9 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static Process WriteText(this Process destination, string text)
   {
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+    if (text is null) throw new ArgumentNullException(nameof(text));
+
     destination.StandardInput.WriteText(text);
     return destination;
   }
@@ -165,7 +180,11 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static async Task<Process> WriteBytesAsync(this Process destination, IEnumerable<byte> bytes, CancellationToken cancellation = default)
   {
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+    if (bytes is null) throw new ArgumentNullException(nameof(bytes));
+
     await destination.StandardInput.BaseStream.WriteBytesAsync(bytes, cancellation).ConfigureAwait(false);
+   
     return destination;
   }
 
@@ -178,7 +197,11 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static async Task<Process> WriteTextAsync(this Process destination, string text, CancellationToken cancellation = default)
   {
+    if (destination is null) throw new ArgumentNullException(nameof(destination));
+    if (text is null) throw new ArgumentNullException(nameof(text));
+
     await destination.StandardInput.WriteTextAsync(text, cancellation).ConfigureAwait(false);
+
     return destination;
   }
 
@@ -190,7 +213,11 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static IEnumerable<byte> WriteTo(this IEnumerable<byte> bytes, Process process)
   {
+    if (bytes is null) throw new ArgumentNullException(nameof(bytes));
+    if (process is null) throw new ArgumentNullException(nameof(process));
+
     process.WriteBytes(bytes);
+
     return bytes;
   }
 
@@ -202,7 +229,11 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static string WriteTo(this string text, Process process)
   {
+    if (text is null) throw new ArgumentNullException(nameof(text));
+    if (process is null) throw new ArgumentNullException(nameof(process));
+
     process.WriteText(text);
+
     return text;
   }
 
@@ -215,7 +246,11 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static async Task<IEnumerable<byte>> WriteToAsync(this IEnumerable<byte> bytes, Process process, CancellationToken cancellation = default)
   {
+    if (bytes is null) throw new ArgumentNullException(nameof(bytes));
+    if (process is null) throw new ArgumentNullException(nameof(process));
+
     await process.WriteBytesAsync(bytes, cancellation).ConfigureAwait(false);
+
     return bytes;
   }
 
@@ -228,7 +263,11 @@ public static class ProcessExtensions
   /// <returns></returns>
   public static async Task<string> WriteToAsync(this string text, Process process, CancellationToken cancellation = default)
   {
+    if (text is null) throw new ArgumentNullException(nameof(text));
+    if (process is null) throw new ArgumentNullException(nameof(process));
+
     await process.WriteTextAsync(text, cancellation).ConfigureAwait(false);
+
     return text;
   }
 }
